refactor(platform-currencies): rename store dir to legacy, add FUTURE.md
Rename src/store -> src/legacy to signal the bridge is temporary.

Add src/legacy/FUTURE.md: why it exists, when it is removed (gated on coin-module decoupling,
owned by another team).

Co-locate test files (drop __tests__ folders). Drop ticket refs from committed files (repo is
public, tickets are not).
