+++ features/platform/currencies/src/legacy/index.ts
@@ -1,0 +1,1 @@
+export * from "./store";
