+++ features/platform/currencies/src/legacy/store/index.ts
