+++ features/platform/currencies/src/renderWithStore.tsx
